feat: configure authentication and authorization

+++ application/controllers/Authors.php
@@ -14,6 +14,7 @@ class Authors extends RestController {
 		parent::__construct();
 
 		$this->load->model('Author');
+		$this->load->model('User');
 	}
 
 	/** HTTP_GET: Get the details of the author for the given book author id
@@ -21,26 +22,42 @@ class Authors extends RestController {
    * @return HTTP_Response The HTTP status code according to the result and the data body
    */
 	public function index_get() {
-		$author_id = $this->input->get('id');
-
-		try {
-			$author_record = $this->Author->get_author($author_id);
-		}
-		catch(AuthorIdRequiredException $ex) {
-			$error_response = json_encode(array('message' => $ex->errorMessage()));
-      $this->response($error_response, RestController::HTTP_BAD_REQUEST);
-		}
-		catch(AuthorDoesNotExistException $ex) {
-			$error_response = json_encode(array('message' => $ex->errorMessage()));
-      $this->response($error_response, RestController::HTTP_BAD_REQUEST);
-		}
-		catch(Exception $ex) {
-      $error_response = json_encode(array('message' => 'Oops! Something went wrong :/'));
-      $this->response($error_response, RestController::HTTP_INTERNAL_ERROR);
-    }
+		$headers = $this->input->request_headers();
+		if (isset($headers['Authorization'])) {
+			$auth_header = $headers['Authorization'];
+			$auth_token = preg_replace('/^Bearer\s*/', '', $auth_header);
+			
+			if ($this->User->is_valid_token($auth_token)) {
+				$author_id = $this->input->get('id');
 		
-		$success_response = json_encode($author_record);
-		$this->response($success_response, RestController::HTTP_OK);
+				try {
+					$author_record = $this->Author->get_author($author_id);
+				}
+				catch(AuthorIdRequiredException $ex) {
+					$error_response = json_encode(array('message' => $ex->errorMessage()));
+					$this->response($error_response, RestController::HTTP_BAD_REQUEST);
+				}
+				catch(AuthorDoesNotExistException $ex) {
+					$error_response = json_encode(array('message' => $ex->errorMessage()));
+					$this->response($error_response, RestController::HTTP_BAD_REQUEST);
+				}
+				catch(Exception $ex) {
+					$error_response = json_encode(array('message' => 'Oops! Something went wrong :/'));
+					$this->response($error_response, RestController::HTTP_INTERNAL_ERROR);
+				}
+				
+				$success_response = json_encode($author_record);
+				$this->response($success_response, RestController::HTTP_OK);
+			}
+			else {
+				$error_response = json_encode(array('status' => 'error', 'message' => 'Unauthorized!'));
+				$this->response($error_response, RestController::HTTP_UNAUTHORIZED);
+			}
+		}
+		else {
+			$error_response = json_encode(array('status' => 'error', 'message' => 'No authorization header found!'));
+			$this->response($error_response, RestController::HTTP_UNAUTHORIZED);
+		}
 	}
 
 	/** HTTP_POST: Create a new book author with the provided name
@@ -48,27 +65,43 @@ class Authors extends RestController {
    * @return HTTP_Response The HTTP status code according to the result and the data body
    */
 	public function index_post() {
-		$author_name = $this->input->get('name');
+		$headers = $this->input->request_headers();
+		if (isset($headers['Authorization'])) {
+			$auth_header = $headers['Authorization'];
+			$auth_token = preg_replace('/^Bearer\s*/', '', $auth_header);
+			
+			if ($this->User->is_valid_token($auth_token)) {
+				$author_name = $this->input->get('name');
+				
+				try {
+					$author_id = $this->Author->create($author_name);
+				}
+				catch(InvalidArgumentException $ex) {
+					$error_response = json_encode(array('message' => 'Invlid or Missing Argument!'));
+					$this->response($error_response, RestController::HTTP_BAD_REQUEST);
+				}
+				catch(Exception $ex) {
+					$error_response = json_encode(array('message' => 'Oops! Something went wrong :/'));
+					$this->response($error_response, RestController::HTTP_INTERNAL_ERROR);
+				}
 		
-		try {
-			$author_id = $this->Author->create($author_name);
+				if(!$author_id) {
+					$error_response = json_encode(array('message' => 'Oops! Something went wrong :/'));
+					$this->response($error_response, RestController::HTTP_INTERNAL_ERROR);
+				}
+		
+				$success_response = json_encode(array('created_author_id' => $author_id));
+				$this->response($success_response, RestController::HTTP_CREATED);
+			}
+			else {
+				$error_response = json_encode(array('status' => 'error', 'message' => 'Unauthorized!'));
+				$this->response($error_response, RestController::HTTP_UNAUTHORIZED);
+			}
 		}
-		catch(InvalidArgumentException $ex) {
-			$error_response = json_encode(array('message' => 'Invlid or Missing Argument!'));
-      $this->response($error_response, RestController::HTTP_BAD_REQUEST);
+		else {
+			$error_response = json_encode(array('status' => 'error', 'message' => 'No authorization header found!'));
+			$this->response($error_response, RestController::HTTP_UNAUTHORIZED);
 		}
-		catch(Exception $ex) {
-      $error_response = json_encode(array('message' => 'Oops! Something went wrong :/'));
-      $this->response($error_response, RestController::HTTP_INTERNAL_ERROR);
-    }
-
-		if(!$author_id) {
-			$error_response = json_encode(array('message' => 'Oops! Something went wrong :/'));
-      $this->response($error_response, RestController::HTTP_INTERNAL_ERROR);
-		}
-
-		$success_response = json_encode(array('created_author_id' => $author_id));
-		$this->response($success_response, RestController::HTTP_CREATED);
 	}
 
 	/** HTTP_PUT: Update the details of the book author with the provided details
@@ -77,35 +110,51 @@ class Authors extends RestController {
    * @return HTTP_Response The HTTP status code according to the result and the data body
    */
 	public function index_put() {
-		$author_id = $this->input->get('id');
-		$new_name = $this->input->get('new_name');
-
-		$data = array();
-    $new_name && $data['name'] = $new_name;
-
-		try {
-			$result = $this->Author->update($author_id, $data);
+		$headers = $this->input->request_headers();
+		if (isset($headers['Authorization'])) {
+			$auth_header = $headers['Authorization'];
+			$auth_token = preg_replace('/^Bearer\s*/', '', $auth_header);
+			
+			if ($this->User->is_valid_token($auth_token)) {
+				$author_id = $this->input->get('id');
+				$new_name = $this->input->get('new_name');
+		
+				$data = array();
+				$new_name && $data['name'] = $new_name;
+		
+				try {
+					$result = $this->Author->update($author_id, $data);
+				}
+				catch(AuthorIdRequiredException $ex) {
+					$error_response = json_encode(array('message' => $ex->errorMessage()));
+					$this->response($error_response, RestController::HTTP_BAD_REQUEST);
+				}
+				catch(AuthorDoesNotExistException $ex) {
+					$error_response = json_encode(array('message' => $ex->errorMessage()));
+					$this->response($error_response, RestController::HTTP_BAD_REQUEST);
+				}
+				catch(Exception $ex) {
+					$error_response = json_encode(array('message' => 'Oops! Something went wrong :/'));
+					$this->response($error_response, RestController::HTTP_INTERNAL_ERROR);
+				}
+		
+				if(!$result) {
+					$error_response = json_encode(array('message' => 'Oops! Something went wrong :/'));
+					$this->response($error_response, RestController::HTTP_INTERNAL_ERROR);
+				}
+		
+				$success_response = json_encode(array('message' => 'Changed the author details successfully!'));
+				$this->response($success_response, RestController::HTTP_OK);
+			}
+			else {
+				$error_response = json_encode(array('status' => 'error', 'message' => 'Unauthorized!'));
+				$this->response($error_response, RestController::HTTP_UNAUTHORIZED);
+			}
 		}
-		catch(AuthorIdRequiredException $ex) {
-			$error_response = json_encode(array('message' => $ex->errorMessage()));
-      $this->response($error_response, RestController::HTTP_BAD_REQUEST);
+		else {
+			$error_response = json_encode(array('status' => 'error', 'message' => 'No authorization header found!'));
+			$this->response($error_response, RestController::HTTP_UNAUTHORIZED);
 		}
-		catch(AuthorDoesNotExistException $ex) {
-			$error_response = json_encode(array('message' => $ex->errorMessage()));
-      $this->response($error_response, RestController::HTTP_BAD_REQUEST);
-		}
-		catch(Exception $ex) {
-      $error_response = json_encode(array('message' => 'Oops! Something went wrong :/'));
-      $this->response($error_response, RestController::HTTP_INTERNAL_ERROR);
-    }
-
-		if(!$result) {
-			$error_response = json_encode(array('message' => 'Oops! Something went wrong :/'));
-      $this->response($error_response, RestController::HTTP_INTERNAL_ERROR);
-		}
-
-		$success_response = json_encode(array('message' => 'Changed the author details successfully!'));
-		$this->response($success_response, RestController::HTTP_OK);
 	}
 
 	/** HTTP_DELETE: Delete the book author for the given id
@@ -113,30 +162,46 @@ class Authors extends RestController {
    * @return HTTP_Response The HTTP status code according to the result and the data body
    */
 	public function index_delete() {
-		$author_id = $this->input->get('id');
-
-		try {
-			$result = $this->Author->delete($author_id);
+		$headers = $this->input->request_headers();
+		if (isset($headers['Authorization'])) {
+			$auth_header = $headers['Authorization'];
+			$auth_token = preg_replace('/^Bearer\s*/', '', $auth_header);
+			
+			if ($this->User->is_valid_token($auth_token)) {
+				$author_id = $this->input->get('id');
+		
+				try {
+					$result = $this->Author->delete($author_id);
+				}
+				catch(AuthorIdRequiredException $ex) {
+					$error_response = json_encode(array('message' => $ex->errorMessage()));
+					$this->response($error_response, RestController::HTTP_BAD_REQUEST);
+				}
+				catch(AuthorDoesNotExistException $ex) {
+					$error_response = json_encode(array('message' => $ex->errorMessage()));
+					$this->response($error_response, RestController::HTTP_BAD_REQUEST);
+				}
+				catch(Exception $ex) {
+					$error_response = json_encode(array('message' => 'Oops! Something went wrong :/'));
+					$this->response($error_response, RestController::HTTP_INTERNAL_ERROR);
+				}
+		
+				if(!$result) {
+					$error_response = json_encode(array('message' => 'Oops! Something went wrong :/'));
+					$this->response($error_response, RestController::HTTP_INTERNAL_ERROR);
+				}
+		
+				$success_response = json_encode(array('message' => 'Deleted the author successfully!'));
+				$this->response($success_response, RestController::HTTP_OK);
+			}
+			else {
+				$error_response = json_encode(array('status' => 'error', 'message' => 'Unauthorized!'));
+				$this->response($error_response, RestController::HTTP_UNAUTHORIZED);
+			}
 		}
-		catch(AuthorIdRequiredException $ex) {
-			$error_response = json_encode(array('message' => $ex->errorMessage()));
-      $this->response($error_response, RestController::HTTP_BAD_REQUEST);
+		else {
+			$error_response = json_encode(array('status' => 'error', 'message' => 'No authorization header found!'));
+			$this->response($error_response, RestController::HTTP_UNAUTHORIZED);
 		}
-		catch(AuthorDoesNotExistException $ex) {
-			$error_response = json_encode(array('message' => $ex->errorMessage()));
-      $this->response($error_response, RestController::HTTP_BAD_REQUEST);
-		}
-		catch(Exception $ex) {
-      $error_response = json_encode(array('message' => 'Oops! Something went wrong :/'));
-      $this->response($error_response, RestController::HTTP_INTERNAL_ERROR);
-    }
-
-		if(!$result) {
-			$error_response = json_encode(array('message' => 'Oops! Something went wrong :/'));
-      $this->response($error_response, RestController::HTTP_INTERNAL_ERROR);
-		}
-
-		$success_response = json_encode(array('message' => 'Deleted the author successfully!'));
-		$this->response($success_response, RestController::HTTP_OK);
 	}
 }
